Extract Bubble constructor/property to separate file
