+++ lib/index.js
@@ -1,32 +1,6 @@
+const Bubble = require('./bubble');
 var canvas  = document.getElementById('gameCanvas');
 var context = canvas.getContext('2d');
-
-function Bubble(x, y, r, eAngle, canvas, context) {
-  this.x       = x;
-  this.y       = y;
-  this.radius  = r;
-  this.sAngle  = 0;
-  this.eAngle  = eAngle;
-  this.canvas  = canvas;
-  this.context = context;
-}
-
-Bubble.prototype = {
-  draw: function() {
-    context.beginPath();
-    context.arc(this.x, this.y, this.radius, this.sAngle, this.eAngle);
-
-    this.context.fillStyle = "magenta";
-    this.context.fill();
-
-    return this;
-  },
-
-  drop: function() {
-    this.y = this.y + 2;
-    return this;
-  }
-};
 
 var bubbles = [];
 
@@ -36,20 +10,25 @@ for (var j = 0; j < 280; j += 23) {
   }
 }
 
-requestAnimationFrame(function gameLoop() {
-  context.clearRect(0, 0, canvas.width, canvas.height);
-
+function drawBubbles() {
   bubbles.forEach(function(bubble) {
     bubble.draw();
-    dropBubble(bubble);
+
+    bindEvents(bubble);
   });
+}
 
-  requestAnimationFrame(gameLoop);
-});
-
-
-function dropBubble(bubble) {
+function bindEvents(bubble) {
+  // i1 add logic so that a single bubble drops on click
+      // i2 drop matching bubbles
   canvas.addEventListener('click', function() {
     bubble.drop();
   });
 }
+
+requestAnimationFrame(function gameLoop() {
+  context.clearRect(0, 0, canvas.width, canvas.height);
+  drawBubbles();
+
+  requestAnimationFrame(gameLoop);
+});
